geometry buffer rename and documentation
