+++ Source/MonoGame.Extended/MonoGame.Extended.csproj
@@ -59,7 +59,7 @@
     <Compile Include="Collections\Deque.cs" />
     <Compile Include="Collections\ListExtensions.cs" />
     <Compile Include="Collections\Pool.cs" />
-    <Compile Include="Graphics\MeshBufferType.cs" />
+    <Compile Include="Graphics\GeometryBufferType.cs" />
     <Compile Include="Shapes\AffineMatrix.cs" />
     <Compile Include="Shapes\ArcF.cs" />
     <Compile Include="Shapes\Line2F.cs" />
@@ -72,11 +72,9 @@
     <Compile Include="Shapes\Triangulation\PlanarTriangleOutputDelegate.cs" />
     <Compile Include="Graphics\Batching\IBatchItemData.cs" />
     <Compile Include="Graphics\Effects\ITextureEffect2D.cs" />
-    <Compile Include="Graphics\MeshBuffer.cs" />
+    <Compile Include="Graphics\GeometryBuffer.cs" />
     <Compile Include="Graphics\MeshBuilder.cs" />
     <Compile Include="Graphics\MeshBuilderColorTextureExtensions.cs" />
-    <Compile Include="Point3F.cs" />
-    <Compile Include="Point2F.cs" />
     <Compile Include="Shapes\Explicit\ShapeBuilder.cs" />
     <Compile Include="Graphics\Effects\EffectResource.cs" />
     <Compile Include="Graphics\Effects\ShapeEffect.cs" />
